Updated Firewall code with frequency processing changes.

+++ FWCoreLibrary/FWCore.h
@@ -32,5 +32,5 @@
 // FW Core API
 FWCORE_API int createFWInstance(const char* pszCfgFile);
 FWCORE_API int destroyFWInstance(int iFwInstance);
-FWCORE_API int processMessage(int iFwInstance, const int iMsgIdx, const unsigned char* pPayload, const int nPayloadSz, const long timestamp);
+FWCORE_API int processMessage(int iFwInstance, const int iMsgIdx, const unsigned char* pPayload, const int nPayloadSz, const unsigned long long timestamp);
 FWCORE_API int getLastError(void);
