+++ FWCoreLibrary/FrequencyProcessor.h
@@ -10,7 +10,7 @@ public:
 	FrequencyProcessor(const char * fileName, BaseLogger * logobj, const int notifierTimer);
 	FrequencyProcessor();
 	~FrequencyProcessor();
-	int processNewID(const int ID, const long timestamp);
+	int processNewID(const int ID, const unsigned long long timestamp);
 	void frequencyNotifier();
 	bool thShouldRun;
 	void initNotifier();
@@ -23,7 +23,7 @@ private:
 	BaseLogger * logobj;
 
 	map<int, tuple<int, int, int> > idFrequencyMap;
-	map<unsigned long, unsigned long> idTimestampMap;
+	map<int, unsigned long long> idTimestampMap;
 	map<int, tuple<int, int, int>>::iterator mapitterator;
 	bool thReady = 0;
 	thread my_thread_;
